requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
dbt-postgres
pytest
pandas
python-dotenv
httpx
ultralytics
opencv-python
fastapi
uvicorn
sqlalchemy
psycopg2-binary
telethon
